# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PortaSplit Stock Alerts API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  model: string;
  slug: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  description?: string | null;
  checkIntervalSeconds?: number;
}

export interface ProductDetail {
  id: number;
  name: string;
  model: string;
  slug: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
  storeCount: number;
  inStockCount: number;
  lowStockCount?: number;
  outOfStockCount?: number;
  reservedCount?: number;
  /** @nullable */
  lastCheckedAt: string | null;
  checkIntervalSeconds?: number;
}

export interface Store {
  id: number;
  name: string;
  chain: string;
  address: string;
  city: string;
  postalCode: string;
  lat: number;
  lng: number;
  /** @nullable */
  url?: string | null;
}

export type StoreWithStockStockStatus = typeof StoreWithStockStockStatus[keyof typeof StoreWithStockStockStatus];


export const StoreWithStockStockStatus = {
  in_stock: 'in_stock',
  low_stock: 'low_stock',
  out_of_stock: 'out_of_stock',
  reserved: 'reserved',
} as const;

export interface StoreWithStock {
  id: number;
  name: string;
  chain: string;
  address: string;
  city: string;
  postalCode: string;
  lat: number;
  lng: number;
  /** @nullable */
  url?: string | null;
  stockStatus: StoreWithStockStockStatus;
  /** @nullable */
  stockQty?: number | null;
  /** @nullable */
  lastCheckedAt: string | null;
  /** @nullable */
  productId?: number | null;
}

export type StockMapPointStatus = typeof StockMapPointStatus[keyof typeof StockMapPointStatus];


export const StockMapPointStatus = {
  in_stock: 'in_stock',
  low_stock: 'low_stock',
  out_of_stock: 'out_of_stock',
  reserved: 'reserved',
} as const;

export interface StockMapPoint {
  storeId: number;
  name: string;
  chain: string;
  lat: number;
  lng: number;
  status: StockMapPointStatus;
  /** @nullable */
  stockQty?: number | null;
  city?: string;
}

export interface StockSummary {
  totalStores: number;
  inStockCount: number;
  lowStockCount: number;
  outOfStockCount: number;
  reservedCount: number;
  /** @nullable */
  lastCheckedAt: string | null;
  activeSubscribers?: number;
}

export interface ChainStat {
  chain: string;
  storeCount: number;
  inStockCount: number;
  /** @nullable */
  logoUrl?: string | null;
}

export interface AlertInput {
  email: string;
  productId: number;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  radiusKm?: number | null;
  /** @nullable */
  chain?: string | null;
  /** @nullable */
  planId?: string | null;
}

export type AlertSubscriptionStatus = typeof AlertSubscriptionStatus[keyof typeof AlertSubscriptionStatus];


export const AlertSubscriptionStatus = {
  pending: 'pending',
  active: 'active',
  cancelled: 'cancelled',
} as const;

export interface AlertSubscription {
  id: number;
  email: string;
  productId: number;
  status: AlertSubscriptionStatus;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  radiusKm?: number | null;
  /** @nullable */
  chain?: string | null;
  /** @nullable */
  planId?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface Plan {
  id: string;
  name: string;
  durationDays: number;
  priceEur: number;
  checkIntervalSeconds: number;
  features: string[];
  popular?: boolean;
  /** @nullable */
  polarProductId?: string | null;
}

export interface CheckoutInput {
  planId: string;
  email: string;
  /** @nullable */
  productId?: number | null;
}

export interface CheckoutResult {
  checkoutUrl: string;
}

export type ListStoresParams = {
productId?: number;
chain?: string;
status?: ListStoresStatus;
postalCode?: string;
radiusKm?: number;
};

export type ListStoresStatus = typeof ListStoresStatus[keyof typeof ListStoresStatus];


export const ListStoresStatus = {
  in_stock: 'in_stock',
  low_stock: 'low_stock',
  out_of_stock: 'out_of_stock',
} as const;

export type GetStockMapParams = {
productId?: number;
};

export type GetStockSummaryParams = {
productId?: number;
};