# lib/api-zod/src/generated/types/alertSubscription.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PortaSplit Stock Alerts API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSubscriptionStatus } from './alertSubscriptionStatus';

export interface AlertSubscription {
  id: number;
  email: string;
  productId: number;
  status: AlertSubscriptionStatus;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  radiusKm?: number | null;
  /** @nullable */
  chain?: string | null;
  /** @nullable */
  planId?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}