# lib/api-zod/src/generated/types/chainStat.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PortaSplit Stock Alerts API
 * OpenAPI spec version: 0.1.0
 */

export interface ChainStat {
  chain: string;
  storeCount: number;
  inStockCount: number;
  /** @nullable */
  logoUrl?: string | null;
}