# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PortaSplit Stock Alerts API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  name: string;
  durationDays: number;
  priceEur: number;
  checkIntervalSeconds: number;
  features: string[];
  popular?: boolean;
  /** @nullable */
  polarProductId?: string | null;
}