# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PortaSplit Stock Alerts API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  model: string;
  slug: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  description?: string | null;
  checkIntervalSeconds?: number;
}