# lib/api-zod/src/generated/types/stockSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PortaSplit Stock Alerts API
 * OpenAPI spec version: 0.1.0
 */

export interface StockSummary {
  totalStores: number;
  inStockCount: number;
  lowStockCount: number;
  outOfStockCount: number;
  reservedCount: number;
  /** @nullable */
  lastCheckedAt: string | null;
  activeSubscribers?: number;
}